requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
web3==7.14.1
mcp==1.27.0
requests==2.33.1
python-dotenv==1.2.2
eth-account==0.13.7
pydantic==2.12.5
pydantic-settings==2.13.1
